Rename catkit/testbed/Experiment.py -> catkit/testbed/experiment.py
