+++ catkit/testbed/__init__.py
@@ -1,3 +1,3 @@
-from catkit.testbed.Experiment import devices
-from catkit.testbed.Experiment import Experiment
-from catkit.testbed.Experiment import SafetyException
+from catkit.testbed.experiment import devices
+from catkit.testbed.experiment import Experiment
+from catkit.testbed.experiment import SafetyException
